+++ src/xscen/io.py
@@ -5,6 +5,7 @@ import json
 import logging
 import os
 import shutil as sh
+import warnings
 from collections import defaultdict
 from collections.abc import Sequence
 from inspect import signature
@@ -433,6 +434,8 @@ def save_to_zarr(  # noqa: C901
     itervar: bool = False,
     timeout_cleanup: bool = True,
     strip_cat_metadata: bool = False,
+    zip_zarrdir: str | None = None,
+    zip_kwargs: dict | None = None,
 ):
     """
     Save a Dataset to Zarr format, rechunking and compressing if requested.
@@ -445,6 +448,7 @@ def save_to_zarr(  # noqa: C901
         The Dataset to be saved.
     filename : str or os.PathLike
         Name of the Zarr file to be saved.
+        If this ends with .zip, the zarr directory will be zipped after saving.
     rechunk : dict, optional
         This is a mapping from dimension name to new chunks (in any format understood by dask).
         Spatial dimensions can be generalized as 'X' and 'Y' which will be mapped to the actual grid type's
@@ -477,6 +481,18 @@ def save_to_zarr(  # noqa: C901
         This does nothing if `compute` is False.
     strip_cat_metadata : bool
         If True (default), strips all catalog-added attributes before saving the dataset.
+    zip_zarrdir: string, optional
+        If given and filename ends in zip, the saved zarr directory is first saved in this directory,
+        then zipped to `filename`. For the initial zarr, if the zip_zarrdir ends in .zarr, it is used as is. If it
+        does not end in .zarr, the name of `filename` is used inside this dir.
+        If given, but `filename` does not end with .zip, this is ignored.
+        If not given and filename ends in zip, the initial zarr is saved directly to `filename` without .zip suffix, then zip to `filename`.
+        It is possible to pass a path with environment variables like ${SLURM_TMPDIR} to ``zip_zarrdir``.
+    zip_kwargs : dict, optional
+        If given and `filename` ends in zip, the saved zarr directory is zipped using ``xs.io.zip_directory(**zip_kwargs)``.
+        If `zipfile` arg is given, it is ignored. The `zipfile` is always set to `filename`.
+        If given but `filename` does not end with .zip, this is ignored.
+        Tip: Pass `delete=True` here to erase the temporary zarr file.
 
     Returns
     -------
@@ -495,6 +511,23 @@ def save_to_zarr(  # noqa: C901
         ds = rechunk_for_saving(ds, rechunk)
 
     path = Path(filename)
+
+    # if path is zip, make sure it is zipped later
+    if path.suffix == ".zip":
+        zip_kwargs = zip_kwargs or {}
+        if "zipfile" in zip_kwargs:
+            warnings.warn("The 'zipfile' argument in zip_kwargs will be ignored since the filename ends with .zip", stacklevel=2)
+        zip_kwargs["zipfile"] = path
+
+        # make path for zarr
+        if zip_zarrdir and Path(zip_zarrdir).suffix == ".zarr":
+            path = Path(zip_zarrdir)
+        elif zip_zarrdir:
+            # expand var allows to pass ${SLURM_TMPDIR} or similar
+            path = Path(os.path.expandvars(zip_zarrdir)) / path.with_suffix("").name
+        else:
+            path = Path(path.with_suffix(""))
+
     path.parent.mkdir(parents=True, exist_ok=True)
     if path.is_dir():
         tgtds = zarr.open(str(path), mode="r")
@@ -503,7 +536,6 @@ def save_to_zarr(  # noqa: C901
 
     if encoding:
         encoding = encoding.copy()
-
     # Prepare to_zarr kwargs
     if zarr_kwargs is None:
         zarr_kwargs = {}
@@ -546,7 +578,6 @@ def save_to_zarr(  # noqa: C901
         # Remove a few problematic entries from encoding, since it can cause issues with some engines.
         ds[var].encoding.pop("original_shape", None)
         ds[var].encoding.pop("dtype", None)
-
     if len(ds.data_vars) == 0:
         return None
 
@@ -556,7 +587,6 @@ def save_to_zarr(  # noqa: C901
     _coerce_attrs(ds.attrs)
     for var in ds.variables.values():
         _coerce_attrs(var.attrs)
-
     if itervar:
         zarr_kwargs["compute"] = True
         allvars = set(ds.data_vars.keys())
@@ -571,7 +601,7 @@ def save_to_zarr(  # noqa: C901
             logger.debug(msg)
             dsvar = ds.drop_vars(allvars - {name})
             try:
-                dsvar.to_zarr(
+                z = dsvar.to_zarr(
                     path,
                     mode="a",
                     encoding={k: v for k, v in (encoding or {}).items() if k in dsvar},
@@ -586,11 +616,16 @@ def save_to_zarr(  # noqa: C901
         msg = f"Writing {list(ds.data_vars.keys())} for {filename}."
         logger.debug(msg)
         try:
-            return ds.to_zarr(filename, compute=compute, mode="a", encoding=encoding, **zarr_kwargs)
+            z = ds.to_zarr(path, compute=compute, mode="a", encoding=encoding, **zarr_kwargs)
         except TimeoutException:
             if timeout_cleanup:
                 clean_incomplete(path, incomplete=list(ds.data_vars.keys()))
             raise
+
+    if Path(filename).suffix == ".zip":
+        zip_directory(path, **zip_kwargs)
+        z = None
+    return z
 
 
 def _to_dataframe(